+++ src/index.ts
@@ -1,5 +1,4 @@
 import { OrgChart } from './OrgChart';
-export * from './utils';
 export * from './types';
 
 export default OrgChart;
