+++ src/options.ts
@@ -1,7 +1,8 @@
 import { d3 } from './constants';
 import { D3Node, OrgChartConnection, OrgChartDataItem, OrgChartOptions } from './types';
 import { BaseType } from 'd3-selection';
-import { diagonal, getTextWidth, hdiagonal } from './utils';
+import { getTextWidth } from './utils/core';
+import { diagonal, hdiagonal } from './utils/calculations';
 
 const canvasContext = document.createElement('canvas').getContext('2d');
